requirements.txt
# === Core FastAPI ===
fastapi==0.115.8
uvicorn==0.34.0
python-multipart==0.0.20
aiofiles==23.2.1
pydantic==1.10.13

# === PDF Manipulation ===
PyPDF2==3.0.1
pymupdf==1.25.5
pdfplumber==0.10.3
pdfminer.six==20221105
reportlab==4.0.9
pdfrw==0.4

# === Image Handling / Rendering ===
pillow==11.1.0
pdf2image==1.17.0
pypdfium2==4.24.0
opencv-python==4.8.0.76
numpy==1.26.3

# === OCR ===
pytesseract==0.3.10

# === Office File Manipulation ===
python-docx==1.1.0
python-pptx==0.6.23
openpyxl==3.1.2

# === Text & Batch Utils ===
tqdm==4.66.2
shutilwhich==1.1.0  # utile pour vérifier si pandoc/libreoffice/tesseract sont installés

# === QR Code / Barcodes ===
qrcode==7.4.2
python-barcode==0.15.1

# === Cryptography / Protection ===
cryptography==41.0.7

# === Packaging (si tu veux compiler l'app plus tard) ===
pyinstaller==5.13.0

# === (Optionnel) Interface Pandoc via Python ===
pandoc==2.4
